+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased (16.0.1)]
+
+### Fixed
+
+- An `sfdp` crash has been fixed. This was a regression in Graphviz 15.1.1.
+  \#2852
+
 ## [16.0.0] – 2026-08-14
 
 ### Added
@@ -3619,6 +3626,7 @@ March 13, 2000: Use AM_PROG_LIBTOOL instead of AC_PROG_LIBTOOL
    in configure.in.  John Ellson <ellson@graphviz.org>
 ```
 
+[Unreleased (16.0.1)]: https://gitlab.com/graphviz/graphviz/compare/16.0.0...main
 [16.0.0]: https://gitlab.com/graphviz/graphviz/compare/15.1.1...16.0.0
 [15.1.1]: https://gitlab.com/graphviz/graphviz/compare/15.1.0...15.1.1
 [15.1.0]: https://gitlab.com/graphviz/graphviz/compare/15.0.0...15.1.0